fix: modify create_record to set id of new records
This fix makes a change to the create_record method to properly send
the record id when we want to deliberately set the UUID of the new
record.  So that existing behaviour is preserved, we simply look for
the id in the record_data provided to the method.  This commit also
adds an integration test of this behaviour.