feat: generalise testing config to work with external test container setup
As reported in https://github.com/IATI/iati-registry-libsuitecrm/issues/4 the
bitnami docker images of SuiteCRM are now deprecated.  This commit changes the
testing context to use an external docker image and removes the current docker
configuration that relies on the bitnami docker image.  Use of an external
configuration, such as the test docker setup in
http://github.com/IATI/iati-suitecrm is used as we need to test that this
works against our actual SuiteCRM codebase.