test: fix to clear SuiteCRM tables used by tests
This commit adds a fixture which runs before each test,
clearing out the tables that are used by the tests to avoid
test failures due stale data being in the tables and polluting
counts or DB requests that assume only a single row will be
returned.
The 'crm' fixture is altered to have function scope, and to
depend on the clear_tables_used_by_tests fixture, which
ensures the instance's OAuth token isn't removed by the
clear up fixture.
Resolves #12.